refactor(navigation): get rid of ion-nav-controller, get nav working correctly in DOM, angular, react
