+++ packages/angular/src/components/ion-nav.ts
@@ -1,17 +1,23 @@
 import {
   ChangeDetectorRef,
   Component,
-  OnInit,
+  ComponentFactoryResolver,
+  ComponentRef,
+  ElementRef,
+  Injector,
   ReflectiveInjector,
+  Type,
   ViewContainerRef,
   ViewChild
 } from '@angular/core';
 
-import { PublicNav } from '@ionic/core';
+import { FrameworkDelegate } from '@ionic/core';
 
 import { getProviders } from '../di/di';
-import { AngularFrameworkDelegate } from '../providers/angular-framework-delegate';
-import { AngularViewController } from '../types/angular-view-controller';
+import { AngularComponentMounter } from '../providers/angular-component-mounter';
+import { AngularMountingData } from '../types/interfaces';
+
+const elementToComponentRefMap = new Map<HTMLElement, ComponentRef<any>>();
 
 @Component({
   selector: 'ion-nav',
@@ -19,41 +25,41 @@ import { AngularViewController } from '../types/angular-view-controller';
     <div #viewport class="ng-nav-viewport"></div>
   `
 })
-export class IonNavDelegate implements OnInit {
+export class IonNavDelegate implements FrameworkDelegate {
 
   @ViewChild('viewport', { read: ViewContainerRef}) viewport: ViewContainerRef;
 
-  constructor(private changeDetection: ChangeDetectorRef, private angularFrameworkDelegate: AngularFrameworkDelegate) {
+  constructor(private elementRef: ElementRef, private changeDetection: ChangeDetectorRef, private angularComponentMounter: AngularComponentMounter, private injector: Injector, private componentResolveFactory: ComponentFactoryResolver) {
+    this.elementRef.nativeElement.delegate = this;
+
   }
 
-  ngOnInit() {
-    const controllerElement = document.querySelector('ion-nav-controller') as any;
-    controllerElement.delegate = this;
+  async attachViewToDom(elementOrContainerToMountTo: HTMLIonNavElement, elementOrComponentToMount: Type<any>,
+                    _propsOrDataObj?: any, _classesToAdd?: string[]): Promise<AngularMountingData> {
+
+    const componentProviders = ReflectiveInjector.resolve(getProviders(elementOrContainerToMountTo));
+    console.log('componentProviders: ', componentProviders);
+
+    const element = document.createElement('ion-page');
+    for (const clazz of _classesToAdd) {
+      element.classList.add(clazz);
+    }
+
+    elementOrContainerToMountTo.appendChild(element);
+    const mountingData = await this.angularComponentMounter.attachViewToDom(element, elementOrComponentToMount, [], this.changeDetection, this.componentResolveFactory, this.injector);
+    mountingData.element = element;
+
+    elementToComponentRefMap.set(mountingData.angularHostElement, mountingData.componentRef);
+
+    return mountingData;
   }
 
-  attachViewToDom(nav: PublicNav, enteringView: AngularViewController): Promise<any> {
-
-    const componentProviders = ReflectiveInjector.resolve(getProviders(nav.element as HTMLIonNavElement));
-    return this.angularFrameworkDelegate.attachViewToDom(enteringView.component, this.viewport, componentProviders, this.changeDetection).then((angularMountingData) => {
-
-      enteringView.componentFactory = angularMountingData.componentFactory;
-      enteringView.injector = angularMountingData.childInjector;
-      enteringView.componentRef = angularMountingData.componentRef;
-      enteringView.instance = angularMountingData.componentRef.instance;
-      enteringView.angularHostElement = angularMountingData.componentRef.location.nativeElement;
-      enteringView.element = angularMountingData.componentRef.location.nativeElement.querySelector('ion-page');
-    });
-  }
-
-  removeViewFromDom(_nav: PublicNav, viewController: AngularViewController) {
-    return this.angularFrameworkDelegate.removeViewFromDom((viewController as any).componentRef).then(() => {
-      viewController.componentFactory = null;
-      viewController.injector = null;
-      viewController.componentRef = null;
-      viewController.instance = null;
-      viewController.angularHostElement = null;
-      viewController.element = null;
-    });
+  async removeViewFromDom(_parentElement: HTMLElement, childElement: HTMLElement) {
+    const componentRef = elementToComponentRefMap.get(childElement);
+    if (componentRef) {
+      return this.angularComponentMounter.removeViewFromDom(componentRef);
+    }
+    return Promise.resolve();
   }
 }
