+++ packages/angular/src/module.ts
@@ -16,7 +16,7 @@ import { IonNavDelegate } from './components/ion-nav';
 /* Providers */
 import { ActionSheetController } from './providers/action-sheet-controller';
 import { AlertController } from './providers/alert-controller';
-import { AngularFrameworkDelegate } from './providers/angular-framework-delegate';
+import { AngularComponentMounter } from './providers/angular-component-mounter';
 import { LoadingController } from './providers/loading-controller';
 import { ToastController } from './providers/toast-controller';
 
@@ -48,7 +48,7 @@ export class IonicAngularModule {
       providers: [
         AlertController,
         ActionSheetController,
-        AngularFrameworkDelegate,
+        AngularComponentMounter,
         LoadingController,
         ToastController
       ]
