--- packages/core/src/components/nav-controller/readme.md
@@ -1,57 +1,0 @@
-# ion-nav-controller
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-#### delegate
-
-any
-
-
-## Attributes
-
-#### delegate
-
-any
-
-
-## Methods
-
-#### insert()
-
-
-#### insertPages()
-
-
-#### pop()
-
-
-#### popTo()
-
-
-#### popToRoot()
-
-
-#### push()
-
-
-#### removeIndex()
-
-
-#### removeView()
-
-
-#### setPages()
-
-
-#### setRoot()
-
-
-
-----------------------------------------------
-
-*Built by [StencilJS](https://stenciljs.com/)*
