+++ packages/core/src/components/nav/nav-interfaces.ts
@@ -1,18 +1,34 @@
+
+/* it is very important to keep this interface in sync with ./nav */
 import {
   Animation,
   AnimationOptions,
+  FrameworkDelegate,
   Nav,
+  NavOptions,
+  PublicViewController,
   ViewController
-} from '../index';
+} from '../../index';
 
-export interface FrameworkDelegate {
+export interface PublicNav {
+  push(component: any, data?: any, opts?: NavOptions): Promise<any>;
+  pop(opts?: NavOptions): Promise<any>;
+  setRoot(component: any, data?: any, opts?: NavOptions): Promise<any>;
+  insert(insertIndex: number, page: any, params?: any, opts?: NavOptions): Promise<any>;
+  insertPages(insertIndex: number, insertPages: any[], opts?: NavOptions): Promise<any>;
+  popToRoot(opts?: NavOptions): Promise<any>;
+  popTo(indexOrViewCtrl: any, opts?: NavOptions): Promise<any>;
+  removeIndex(startIndex: number, removeCount?: number, opts?: NavOptions): Promise<any>;
+  removeView(viewController: PublicViewController, opts?: NavOptions): Promise<any>;
+  setPages(componentDataPairs: any[], opts?: NavOptions): Promise<any>;
 
-  attachViewToDom(elementOrContainerToMountTo: any, elementOrComponentToMount: any, propsOrDataObj?: any, classesToAdd?: string[]): Promise<FrameworkMountingData>;
-  removeViewFromDom(elementOrContainerToUnmountFrom: any, elementOrComponentToUnmount: any): Promise<FrameworkMountingData>;
-}
+  getActive?(): PublicViewController;
+  getPrevious?(view?: PublicViewController): PublicViewController;
+  canGoBack?(): boolean;
+  canSwipeBack?(): boolean;
+  getFirstView?(): PublicViewController;
 
-export interface FrameworkMountingData {
-  element: HTMLElement;
+  element?: HTMLElement;
 }
 
 export interface NavContainer {
