+++ packages/core/src/utils/dom-framework-delegate.ts
@@ -1,9 +1,9 @@
-import { FrameworkDelegate, FrameworkMountingData, } from '../../index';
-import { isString } from '../../utils/helpers';
+import { FrameworkDelegate, FrameworkMountingData, } from '../index';
+import { isString } from './helpers';
 
 export class DomFrameworkDelegate implements FrameworkDelegate {
-  attachViewToDom(parentElement: HTMLElement, tagOrElement: string | HTMLElement, propsOrDataObj: any = {}, classesToAdd: string[] = []): Promise<FrameworkMountingData> {
 
+  attachViewToDom(parentElement: HTMLElement, tagOrElement: string | HTMLElement, propsOrDataObj: any = {}, classesToAdd: string[] = []): Promise<FrameworkMountingData> {
     return new Promise((resolve) => {
       const usersElement = (isString(tagOrElement) ? document.createElement(tagOrElement) : tagOrElement) as HTMLElement;
 
@@ -21,7 +21,6 @@ export class DomFrameworkDelegate implements FrameworkDelegate {
   }
 
   removeViewFromDom(parentElement: HTMLElement, childElement: HTMLElement): Promise<FrameworkMountingData> {
-
     parentElement.removeChild(childElement);
     return Promise.resolve({
       element: null
