+++ packages/core/stencil.config.js
@@ -38,7 +38,7 @@ exports.config = {
     { components: ['ion-range', 'ion-range-knob']},
     { components: ['ion-tabs', 'ion-tab', 'ion-tabbar', 'ion-tab-button', 'ion-tab-highlight'] },
     { components: ['ion-toggle'] },
-    { components: ['ion-nav', 'ion-nav-controller'] },
+    { components: ['ion-nav'] },
     { components: ['ion-toast', 'ion-toast-controller'] },
   ],
   collections: [
