+++ packages/demos/angular/src/app/app-routing.module.ts
@@ -9,7 +9,8 @@ const routes: Routes = [
   { path: 'alert', loadChildren: 'app/alert/alert.module#AlertModule' },
   { path: 'actionSheet', loadChildren: 'app/action-sheet/action-sheet.module#ActionSheetModule' },
   { path: 'toast', loadChildren: 'app/toast/toast.module#ToastModule' },
-  { path: 'loading', loadChildren: 'app/loading/loading.module#LoadingModule' }
+  { path: 'loading', loadChildren: 'app/loading/loading.module#LoadingModule' },
+  { path: 'nav', loadChildren: 'app/nav/nav.module#NavModule' }
 ];
 
 @NgModule({
