+++ packages/demos/angular/src/app/home-page/home-page.component.html
@@ -21,4 +21,7 @@
   <li>
     <a href='loading'>Loading Page</a>
   </li>
+  <li>
+    <a href='nav'>Nav Page</a>
+  </li>
 </ul>
