+++ packages/demos/angular/tslint.json
@@ -1,6 +1,6 @@
 {
   "rulesDirectory": [
-    "node_modules/codelyzer"
+    // "node_modules/codelyzer"
   ],
   "rules": {
     "arrow-return-shorthand": true,
